requirements.txt
mss>=10.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
deep-translator>=1.11.0
numpy>=1.24.0
